+++ packages/charting/CHANGELOG.md
@@ -1,8 +1,44 @@
 # Change Log - @uifabric/charting
 
-This log was last generated on Thu, 16 Dec 2021 12:24:57 GMT and should not be manually modified.
+This log was last generated on Fri, 21 Jan 2022 12:28:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [4.20.1](https://github.com/microsoft/fluentui/tree/@uifabric/charting_v4.20.1)
+
+Fri, 21 Jan 2022 12:28:47 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/charting_v4.20.0..@uifabric/charting_v4.20.1)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.181.0 ([PR #21161](https://github.com/microsoft/fluentui/pull/21161) by tmichon@microsoft.com)
+
+## [4.20.0](https://github.com/microsoft/fluentui/tree/@uifabric/charting_v4.20.0)
+
+Fri, 14 Jan 2022 12:22:39 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/charting_v4.19.9..@uifabric/charting_v4.20.0)
+
+### Minor changes
+
+- style control prop added in area chart ([PR #21260](https://github.com/microsoft/fluentui/pull/21260) by v-scharde@microsoft.com)
+
+## [4.19.9](https://github.com/microsoft/fluentui/tree/@uifabric/charting_v4.19.9)
+
+Tue, 11 Jan 2022 12:22:43 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/charting_v4.19.8..@uifabric/charting_v4.19.9)
+
+### Patches
+
+- Make last point in line chart focusable even when hidenonactivedots is true. Also change draw order of points so aren't drawn over other lines. ([PR #21233](https://github.com/microsoft/fluentui/pull/21233) by nojittip@microsoft.com)
+
+## [4.19.8](https://github.com/microsoft/fluentui/tree/@uifabric/charting_v4.19.8)
+
+Wed, 05 Jan 2022 12:29:16 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/charting_v4.19.7..@uifabric/charting_v4.19.8)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.180.3 ([PR #20839](https://github.com/microsoft/fluentui/pull/20839) by tkrasniqi@microsoft.com)
 
 ## [4.19.7](https://github.com/microsoft/fluentui/tree/@uifabric/charting_v4.19.7)