+++ packages/example-app-base/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @uifabric/example-app-base
 
-This log was last generated on Thu, 16 Dec 2021 12:24:57 GMT and should not be manually modified.
+This log was last generated on Fri, 21 Jan 2022 12:28:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [7.21.12](https://github.com/microsoft/fluentui/tree/@uifabric/example-app-base_v7.21.12)
+
+Fri, 21 Jan 2022 12:28:47 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/example-app-base_v7.21.11..@uifabric/example-app-base_v7.21.12)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.181.0 ([PR #21161](https://github.com/microsoft/fluentui/pull/21161) by tmichon@microsoft.com)
+
+## [7.21.11](https://github.com/microsoft/fluentui/tree/@uifabric/example-app-base_v7.21.11)
+
+Wed, 05 Jan 2022 12:29:16 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/example-app-base_v7.21.10..@uifabric/example-app-base_v7.21.11)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.180.3 ([PR #20839](https://github.com/microsoft/fluentui/pull/20839) by tkrasniqi@microsoft.com)
 
 ## [7.21.10](https://github.com/microsoft/fluentui/tree/@uifabric/example-app-base_v7.21.10)