+++ packages/mdl2-theme/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @uifabric/mdl2-theme
 
-This log was last generated on Thu, 16 Dec 2021 12:24:57 GMT and should not be manually modified.
+This log was last generated on Fri, 21 Jan 2022 12:28:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.6.12](https://github.com/microsoft/fluentui/tree/@uifabric/mdl2-theme_v0.6.12)
+
+Fri, 21 Jan 2022 12:28:47 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/mdl2-theme_v0.6.11..@uifabric/mdl2-theme_v0.6.12)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.181.0 ([PR #21161](https://github.com/microsoft/fluentui/pull/21161) by tmichon@microsoft.com)
+
+## [0.6.11](https://github.com/microsoft/fluentui/tree/@uifabric/mdl2-theme_v0.6.11)
+
+Wed, 05 Jan 2022 12:29:16 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/mdl2-theme_v0.6.10..@uifabric/mdl2-theme_v0.6.11)
+
+### Patches
+
+- Bump office-ui-fabric-react to v7.180.3 ([PR #20839](https://github.com/microsoft/fluentui/pull/20839) by tkrasniqi@microsoft.com)
 
 ## [0.6.10](https://github.com/microsoft/fluentui/tree/@uifabric/mdl2-theme_v0.6.10)